requirements.txt
astroid==2.2.5
dj-database-url==0.5.0
Django==2.2.3
django-filter==2.2.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.1
Faker==2.0.0
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
psycopg2==2.8.3
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
whitenoise==4.1.3
wrapt==1.11.2
